backend/requirements.txt
uvicorn
fastapi
numpy
sentence-transformers
faiss-cpu
psycopg2-binary
passlib
bcrypt